Struct stack_dst::StackA

pub struct StackA<T: ?Sized, D: DataBuf> { /* fields omitted */ }

A fixed-capacity stack that can contain dynamically-sized types

Uses an array of usize as a backing store for a First-In, Last-Out stack of items that can unsize to T.

Note: Each item in the stack takes at least one usize (to store the metadata)

Methods

impl<T: ?Sized, D: DataBuf> StackA<T, D>

fn new() -> StackA<T, D>

Construct a new (empty) stack

fn is_empty(&self) -> bool

Tests if the stack is empty

fn push<U: Unsize<T>>(&mut self, v: U) -> Result<(), U>

Push a value at the top of the stack

fn top(&self) -> Option<&T>

Returns a pointer to the top item on the stack

fn top_mut(&mut self) -> Option<&mut T>

Returns a pointer to the top item on the stack (unique/mutable)

fn pop(&mut self)

Pop the top item off the stack

impl<D: DataBuf> StackA<str, D>

fn push_str(&mut self, v: &str) -> Result<(), ()>

Push the contents of a string slice as an item onto the stack

impl<D: DataBuf, T: Clone> StackA<[T], D>

fn push_cloned(&mut self, v: &[T]) -> Result<(), ()>

Pushes a set of items (cloning out of the input slice)